Split Dots components into a separate file
